pyproject.toml
[project]
name = "travel-agent"
version = "0.1.0"
description = "Minimal FastAPI app derived from shtabi-backend"
authors = [{ name = "Shtabi Team" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "langchain>=1.0.4",
    "langchain-openai>=1.0.2",
    "langchain-mcp-adapters>=0.1.0",
    "fastmcp[langchain]>=2.13.0.2",
    "pydantic-settings>=2.0.0",
    "httpx>=0.28.0",
    "openai>=1.51.0",
    "groq>=0.5.0",
    "tavily-python>=0.7.10",
    "python-dotenv>=1.0.1",
    "streamlit>=1.39.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
